+++ src/backend/public_request/util/get_public_request_uuid.hh
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018-2022  CZ.NIC, z. s. p. o.
+ * Copyright (C) 2022  CZ.NIC, z. s. p. o.
  *
  * This file is part of FRED.
  *
@@ -16,23 +16,23 @@
  * You should have received a copy of the GNU General Public License
  * along with FRED.  If not, see <https://www.gnu.org/licenses/>.
  */
-#ifndef AUTHINFO_HH_F36A7EC6F242440990E2AB46C3F298BC
-#define AUTHINFO_HH_F36A7EC6F242440990E2AB46C3F298BC
 
-#include "libfred/public_request/public_request_type_iface.hh"
-#include "src/bin/cli/messenger_params.hh"
+#ifndef GET_PUBLIC_REQUEST_UUID_HH_89C0714FE8E74B319CF72B5234B72D43
+#define GET_PUBLIC_REQUEST_UUID_HH_89C0714FE8E74B319CF72B5234B72D43
+
+#include "libfred/object_state/typedefs.hh"
+#include "libfred/opcontext.hh"
 
 namespace Fred {
 namespace Backend {
 namespace PublicRequest {
-namespace Process {
+namespace Util {
 
-void process_public_request_authinfo_resolved(
-        unsigned long long _public_request_id,
-        const LibFred::PublicRequestTypeIface& _public_request_type,
-        const MessengerArgs& _messenger_args);
+boost::uuids::uuid get_public_request_uuid(
+        LibFred::OperationContext& _ctx,
+        const LibFred::PublicRequestId& _public_request_id);
 
-} // namespace Fred::Backend::PublicRequest::Process
+} // namespace Fred::Backend::PublicRequest::Util
 } // namespace Fred::Backend::PublicRequest
 } // namespace Fred::Backend
 } // namespace Fred